[FIX] : Fix structure
